@@ -14,7 +14,7 @@
 		}
 	],
 	"require": {
-		"php": ">=7.3",
+		"php": ">=7.4",
 		"ext-gd": "*",
 		"ext-json": "*"
 	},
@@ -22,9 +22,9 @@
 		"ext-xdebug": "*",
 		"ergebnis/composer-normalize": "^2.0",
 		"natanfelles/coding-standard": "^1.0",
-		"phpmd/phpmd": "^2.7",
-		"phpstan/phpstan": "^0.11",
-		"phpunit/phpunit": "^8.3"
+		"phpmd/phpmd": "^2.8",
+		"phpstan/phpstan": "^0.12",
+		"phpunit/phpunit": "^8.5"
 	},
 	"autoload": {
 		"psr-4": {
